requirements.txt
PyOpenGL==3.1.0
PyQt3D==5.8
PyQt5>=5.8.1<5.8.2
